+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [KMP] Migration core:log
 - [Feature] Add ready app updates popup and notification dot
 - [Feature] Per app loading in installed tab
+- [Feature] Add ability to preview screenshots of fap entries
 - [Refactor] Migrate to markdown renderer from upstream
 - [Refactor] Basic implementation of new transport ble 
 - [FIX] Replace decompose push to move safety push